Move styling files, working Login container
