+++ static/src/components/App/index.jsx
@@ -1,28 +1,17 @@
 import React from 'react';
-import { Provider } from 'react-redux';
-import configureStore from '../../store';
-import Map from '../Map';
-import Login from '../Login';
-import Logo from '../Logo';
-import Modal from '../Modal';
-import Sidebar from '../Sidebar';
-import MessagesBox from '../MessagesBox';
-import '../../stylesheets/App.scss';
-
-const store = configureStore();
+import LoginContainer from '../../containers/Login/LoginContainer';
+// import Map from '../Map';
+// import Logo from '../Logo';
+// import Modal from '../Modal';
+// import Sidebar from '../Sidebar';
+// import MessagesBox from '../MessagesBox';
+import './App.scss';
 
 function App() {
   return (
-    <Provider store={store}>
-      <div className="app-container">
-        <Login />
-        <Map />
-        <Sidebar />
-        <MessagesBox />
-        <Logo />
-        <Modal />
-      </div>
-    </Provider>
+    <div>
+      <LoginContainer />
+    </div>
   );
 }
