+++ static/src/components/Input/CheckBox.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 $box-border-size: 3px;
