+++ static/src/components/Input/InputText.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 .base-focus {
   box-sizing: border-box;
