+++ static/src/components/Input/OptionsList.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 @import "./InputText";
 
 ol.options-list {
