+++ static/src/components/Input/SelectWithLabel.scss
@@ -1,5 +1,5 @@
 @import "./InputText";
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 .select-with-label {
   margin: 15px 0 30px;
