+++ static/src/components/Input/SliderRange.scss
@@ -1,5 +1,5 @@
 @import "./InputText.scss";
-@import "./range.scss";
+@import "../../stylesheets/range.scss";
 
 .slider-range {
   input[type=range] {
