+++ static/src/components/Logo/Logo.scss
