+++ static/src/components/MIDI/Midi.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 .midi-indicator {
   .selectors {
