+++ static/src/components/Map/Map.scss
@@ -1,5 +1,5 @@
-@import "./variables.json";
-@import "./animations";
+@import "../../stylesheets/variables.json";
+@import "../../stylesheets/animations";
 
 .map-container, svg.map {
   height: 100%;
