+++ static/src/components/MessagesBox/MessagesBox.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 .message-box {
   position: fixed;
