+++ static/src/components/Metronome/Metronome.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 .metronome-wrapper {
   margin-top: 20px;
