+++ static/src/components/Modal/Modal.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 .modal-container {
   position: absolute;
