+++ static/src/components/Paths/Paths.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 .paths-list {
   user-select: none;
