+++ static/src/components/Sidebar/QueryBox.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 .query-form {
 	width: 100%;
