+++ static/src/components/Sidebar/Sidebar.scss
@@ -1,5 +1,5 @@
-@import "./variables.json";
-@import "./button";
+@import "../../stylesheets/variables.json";
+@import "../../stylesheets/button";
 
 .sidebar {
   position: absolute;
