+++ static/src/components/Sidebar/SpacesThumbnails.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 ul.spaces-thumbnails-container {
   list-style: none;
