+++ static/src/components/SoundInfo/SoundInfo.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 $sound-info-modal-width: 400px;
 $modal-transition-duration: 0.3s;
