+++ static/src/components/SoundInfo/Waveform.scss
@@ -1,4 +1,4 @@
-@import "./variables.json";
+@import "../../stylesheets/variables.json";
 
 svg.waveform {
   width: $waveformWidth;
